requirements.txt
psycopg2-binary
pydantic-settings
fastapi
python-multipart
uvicorn[standard]
httpx
Pillow
